fix(function): resolve requires relative to process (#316)
* fix(function): resolve requires relative to process

* fix: ensure subprocess exists

* test: rewrite assertions

+++ packages/function/src/index.js
@@ -62,7 +62,7 @@ module.exports = (
     // main
     const result = await pTimeout(task(), timeout, () => {
       isRejected = true
-      subprocess.kill('SIGKILL')
+      if (subprocess) subprocess.kill('SIGKILL')
       throw browserTimeout({ timeout })
     })
